update route flow

+++ src/api/private/branches/Branch.controller.ts
@@ -4,11 +4,17 @@ import { BranchService } from './Branch.service';
 import { ApiResponse, ApiResponseData, ApiResponseList } from '@/lib/dtos/ApiResponse.dto';
 import { messages } from '@/lib/constants/messages';
 import { Meta } from '@/lib/dtos/Meta.dto';
+import { JwtPayload } from '@/lib/dtos/JwtPayload.dto';
+import { RoleEnum } from '@/lib/enums/RoleEnum';
+import { checkPermissions } from '@/lib/utils/checkPermissions';
 
 const BranchController = honoApp()
 
 BranchController.openapi(BranchListRoute, async (context) => {
   const query = context.req.valid('query')
+  const jwtPayload = new JwtPayload(context.get('jwtPayload'))
+  checkPermissions(jwtPayload, [RoleEnum.Admin, RoleEnum.SuperAdmin])
+
   const [branches, totalData] = await Promise.all([
     BranchService.list(query),
     BranchService.count(query),
@@ -28,6 +34,9 @@ BranchController.openapi(BranchListRoute, async (context) => {
 
 BranchController.openapi(BranchDetailRoute, async (context) => {
   const param = context.req.valid('param');
+  const jwtPayload = new JwtPayload(context.get('jwtPayload'))
+  checkPermissions(jwtPayload, [RoleEnum.Admin, RoleEnum.SuperAdmin])
+
   const branch = await BranchService.get(+param.id);
 
   return context.json(new ApiResponseData({
@@ -39,6 +48,9 @@ BranchController.openapi(BranchDetailRoute, async (context) => {
 
 BranchController.openapi(BranchCreateRoute, async (context) => {
   const payload = context.req.valid('json');
+  const jwtPayload = new JwtPayload(context.get('jwtPayload'))
+  checkPermissions(jwtPayload, [RoleEnum.SuperAdmin])
+
   const branch = await BranchService.create(payload)
 
   return context.json(new ApiResponseData({
@@ -51,6 +63,9 @@ BranchController.openapi(BranchCreateRoute, async (context) => {
 BranchController.openapi(BranchUpdateRoute, async (context) => {
   const param = context.req.valid('param')
   const payload = context.req.valid('json')
+  const jwtPayload = new JwtPayload(context.get('jwtPayload'))
+  checkPermissions(jwtPayload, [RoleEnum.Admin, RoleEnum.SuperAdmin])
+
   const branch = await BranchService.update(+param.id, payload);
 
   return context.json(new ApiResponseData({
@@ -62,6 +77,9 @@ BranchController.openapi(BranchUpdateRoute, async (context) => {
 
 BranchController.openapi(BranchDeleteRoute, async (context) => {
   const param = context.req.valid('param')
+  const jwtPayload = new JwtPayload(context.get('jwtPayload'))
+  checkPermissions(jwtPayload, [RoleEnum.SuperAdmin])
+
   await BranchService.delete(+param.id)
 
   return context.json(new ApiResponse({
